A terrible prototype of a randstring operator
To use: (( randstring(16) )), where 16 is the number of chars required.

The random string generation algorithm is terrible, and if we were to use this
to generate passwords it would also need to be aware of different hashing
algorithms to ensure that the hashes were valid (bcrypt has a header, for
example). Should possibly be renamed randpassword and only take a type of
password to generate.

Absolutely no error handling, will explode violently if you don't provide an
amount. May also explode at any other time for any reason.